Code::Blocks OSX 10.10
Code::Blocks on OSX 10.10 had a hell of a time compiling the source.
Here are the changes I had to make in order for it to compile.

OISPrereqs.h fix requires a compiler flag. Without this, the code
thinks it’s being compiled on GCC with Linux (When OSX uses Clang…)

Everything but the OISInputManager.h fix is hackish. Better solutions
could probably be found.