Fix broken keycodes on Linux
PR #1 did fix the disparency between the alpha key inputs and the
keycode beween X11 and all of the other backends, but it also did break
half of the special characters.

The Keymap has been restored, but only for some characters that we
explicitly convert from Xorg keysym to OIS::KeyCode.

This is an hybrid solution between before and after the "linux keyboard
fix" (that was hal a breakage :wink:)

Signed-off by: Arthur Brainville (Ybalrid) <ybalrid@ybalrid.info>