CMakeLists.txt: Respect user's libdir
Without this patch, on systems where /usr/lib64/ is used
(e.g. Gentoo), libraries are installed
to /usr/lib/ despite the standard CMake arguments
being passed.

It's not exactly clear what *the* canonical way
is to take libdir parameters in CMake, but
GNUInstallDirs mostly handles it for the *NIX case.

To be minimally invasive here, I've only changed
the *NIX path and aligned as much of the logic as possible.

Signed-off-by: Sam James <sam@gentoo.org>